# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apartment Management API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/contract/create": {
            "post": {
                "tags": [
                    "Contract"
                ],
                "summary": "Add a new contract",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "resident_id": {
                                        "type": "integer"
                                    },
                                    "apartment_id": {
                                        "type": "integer"
                                    },
                                    "expired_at": {
                                        "type": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add a new contract"
                    }
                }
            }
        },
        "/api/contract/{id}/update": {
            "put": {
                "tags": [
                    "Contract"
                ],
                "summary": "Update a contract",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "resident_id": {
                                        "type": "integer"
                                    },
                                    "apartment_id": {
                                        "type": "integer"
                                    },
                                    "expired_at": {
                                        "type": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a contract"
                    }
                }
            }
        },
        "/api/contract/{id}/delete": {
            "delete": {
                "tags": [
                    "Contract"
                ],
                "summary": "Delete a contract",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a contract."
                    }
                }
            }
        },
...
        "/api/resident/{id}": {
            "get": {
                "tags": [
                    "Resident"
                ],
                "summary": "Get a detail resident",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a resident."
                    }
                }
            }
        },
        "/api/resident/create": {
            "post": {
                "tags": [
                    "Resident"
                ],
                "summary": "Add a new resident",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "date"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "identity_card_number": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add a new resident"
                    }
                }
            }
        },
        "/api/resident/{id}/update": {
            "put": {
                "tags": [
                    "Resident"
                ],
                "summary": "Update a resident",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "date"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "identity_card_number": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a resident"
                    }
                }
            }
        },
        "/api/resident/{id}/delete": {
            "delete": {
                "tags": [
                    "Resident"
                ],
                "summary": "Delete a resident",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete a resident."
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}